+++ CHANGELOG.md
@@ -7,7 +7,9 @@ rule (it is breaking changed by default).
 
 ## 129.0.0 - 2020-09-29
 
+- Added: `jsdoc-typescript` preset.
 - Update `ecmaVersion` to `2021`
+- Changed: disabled `class-methods-use-this` rule.
 - Changed: minimum required `eslint` version is now `7.10.0`.
 - Changed: minimum required `eslint-plugin-import` version is now `2.22.1`.
 - Changed: minimum required `eslint-plugin-jest` version is now `24.0.0`.